+++ services/auth/errors.go
@@ -1,4 +1,4 @@
-package permissions
+package auth
 
 import (
 	"fmt"
